+++ src/SchemaOrgBuilderServiceProvider.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Trinityrank\LaravelSchemaOrgBuilder;
+
+use Spatie\LaravelPackageTools\Package;
+use Spatie\LaravelPackageTools\PackageServiceProvider;
+
+class LaravelSchemaOrgBuilderServiceProvider extends PackageServiceProvider
+{
+    public function configurePackage(Package $package): void
+    {
+        /*
+         * This class is a Package Service Provider
+         *
+         * More info: https://github.com/spatie/laravel-package-tools
+         */
+        $package
+            ->name('laravel-schema-org-builder')
+            ->hasConfigFile('schema');
+    }
+}